Move API endpoint handlers to api_router
